+++ packages/webrun-site-host/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @statewalker/webrun-site-host
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @statewalker/webrun-site-builder@0.1.2
+
+## 0.1.4
+
+### Patch Changes
+
+- @statewalker/webrun-http-browser@0.4.2
+
+## 0.1.3
+
+### Patch Changes
+
+- @statewalker/webrun-http-browser@0.4.1
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [2291ab3]
+- Updated dependencies [ff650fc]
+- Updated dependencies [c6dc18d]
+  - @statewalker/webrun-http-browser@0.4.0
+
 ## 0.1.1
 
 ### Patch Changes